Drills/src/main/java/com/mycompany/drills/Conditional/MischeviousChildren.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package com.mycompany.drills.Conditional;

/**
 *
 * @author Jesse
 */
//We have two children, a and b, and the parameters aSmile and bSmile indicate if 
//each is smiling. We are in trouble if they are both smiling or if neither of 
//them is smiling. Return true if we are in trouble. 
//
//areWeInTrouble(true, true) -> true
//areWeInTrouble(false, false) -> true
//areWeInTrouble(true, false) -> false
public class MischeviousChildren {

    public boolean areWeInTrouble(boolean aSmile, boolean bSmile) {
      if (aSmile && bSmile || !aSmile && !bSmile)
          return true;
      else return false;
    }

}
